# Prover/perf opt serde