This PR implements issue(s) #

**Runtime and Memory Optimizations**
    * Skip the buffer and the extra copy in `cbor.go` by using EncMode.Marshal
    * Assign the packed payload directly and encode PackedObject only once rather than twice.
    * UnpackStructObject: use field by index (faster, avoids name lookup).  
   
**Refactoring**
  * Moved function to dump perf. logs to csv file from `wizard` pkg to `profiling` pkg
     
**Correctness**
   * Verified through `TestSerdeDW`.
   * Rerun setup for `execution-limitless` prover assets and successfully generated the proof for sepolia test data. Overall proof time ~38min. 

**Results**
Benchmarking (`benchstat`) results are as follows:

**Runtime (sec/op)**
The optimization reduced runtime across all 46 test cases, with an overall geometric mean improvement of 11.32%. Key highlights include:
* SerdeDW/CompiledGL-7: Reduced from 96.04 s to 29.69 s (-69.09%).
* SerdeDW/CompiledGL-2: Reduced from 52.45 s to 36.02 s (-31.33%).
* SerdeDW/CompiledConglomeration: Reduced from 165.0 s to 140.9 s (-14.61%).

**Memory Allocation (B/op)**
Memory allocation improved significantly, with a geometric mean reduction of 27.44% (from 16.47 GiB to 11.95 GiB). Notable reductions include:
* SerdeDW/CompiledGL-6: Reduced from 446.4 GiB to 281.7 GiB (-36.90%).
* SerdeDW/CompiledConglomeration: Reduced from 235.3 GiB to 153.4 GiB (-34.83%).
* SerdeDW/GLModule-2: Reduced from 50.11 GiB to 30.02 GiB (-40.08%).

Profiling data also indicates the overall runtime (serialization + deserialization time) reduction of ~18%. In some cases around 30%.

**Note**: Deployment will also require uploading new `prover-assets` for `execution-limitless` circuit to S3 and EFS.  The newly introduced  optimizations are not backward-compatible (existing serialized assets cannot be deserialized with this update).